pyproject.toml
[tool.poetry]
name = "vim_colorscheme_generator"
version = "0.1.0"
description = ""
authors = ["cohama <cohama@live.jp>"]

[tool.poetry.dependencies]
python = "~3.10.6"
numpy = "^1.24.0"
tensorflow-cpu = "^2.11.0"
tensorflow-datasets = "^4.8.0"
jaxlib = "^0.4.1"
jax = "^0.4.1"
flax = "^0.6.3"
optax = "^0.1.4"
pandas = "^1.5.3"
pillow = "^9.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^5.2"
ipython = "^8.7.0"
pdbpp = "^0.10.3"
mypy = "^0.991"
ruff = "^0.0.221"
isort = "^5.11.4"
black = "^22.12.0"
autoflake = "^2.0.0"

[tool.mypy]
ignore_missing_imports = true
allow_redefinition = true
check_untyped_defs = true

[tool.ruff]
select = ["ALL"]
line-length = 120
ignore = [
  "D100",
  "D101",
  "D103",
  "D104",
  "D415",
  "D400",
  "D102",
  "D107",
  "D212",
  "D202",
  "D403",
  "D105",
  "ANN",
  "FBT",
  "T201",
  "RUF002",
  "DTZ005",
  "PD901",
  "UP015",
  "RET504",
]

[tool.ruff.pydocstyle]
convention = "google"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
